refactor(backends): move Request to lib/backends/
This is preparation for adding additional "backends", like a fetch-based backend
or a `@kubernetes/client-node` one.
