+++ lib/backends/request.test.js
@@ -8,7 +8,7 @@ const Request = require('./request')
 
 const url = 'http://mock.kube.api'
 
-describe('lib.request', () => {
+describe('lib.backends.request', () => {
   describe('Request', () => {
     it('handles empty responses', done => {
       nock(url)
