+++ lib/swagger-client.js
@@ -29,7 +29,7 @@ const path = require('path')
 const zlib = require('zlib')
 
 const getAliases = require('./common').getAliases
-const Request = require('./request')
+const Request = require('./backends/request')
 
 class Root extends Component {
   _getSpec (pathname) {
